Separate out File operations
1. Separate out File operations into their own module
2. Also fix key binding bug - by default Ctrl-O adds a blank line
to a Text widget. We have bound it to Open File in our Text
widget, but the class binding was still being called, so doing
Open File also added a blank line (at the end of the file) so the
file appeared "edited" immediately after loading. By returning
the string "break" from the instance callback, it stops the
class callback being called.