Add "goto" and "see img" buttons
Status bar labels changed to buttons, to give more modern look,
and because buttons change appearance when mouse hovers
over them, rather than coding it, and because they are buttons!

Improved `get_current_image_name` to cope with cursor being
coincident with one or more page marks. Since typically they
are at the start of a line (at least initially) it can be confusing to
place the cursor at the start of a line and the page number
being reported as the previous page.