Check if save required before loading recent file
The "do you want to save" check was missing when loading
from the recent files menu.

Made loading routine `file.py` (which has the check) cope with
being given a filename or not.

Also fixed bug in `Load Image` in view menu - now does the
same as the statusbar `See Img` button.