(Mac) Remove redundant "Quit" item from File menu (#80)
In Mac applications, the standard location for the "Quit" menu item is
under the so-called "Application" menu. Tkinter automatically supplies
this by default. The second Quit item in the file menu is redundant.

The default Quit menu item binds Cmd-Q, so the menu item removed by this
commit was not actually able to bind Cmd-Q, though it clearly attempted
to do so.

I tested both of the Quit menu items. Both behaved in the same way,
asking if the user wished to save their changes before exit (if the file
state was dirty). As far as I can tell, this removal doesn't create any
behavioral change.

Label also changed from "Quit" to "Exit" per request.