Allow filename to be given on command line
1. Parse command line args
2. Allow `guiguts -h`
3. Allow `guiguts file.txt` to load file.txt
4. Allow `guiguts -r 1` or `guiguts --recent 9` to load most
nth most recent file (up to 9)

Revealed a known tkinter issue where if a dialog is popped
before `update_idletasks` is called, focus will not return to
the main text window. Fixed, since this commit can cause
that to happen if the filename given on the command line
does not exist.

Reversed the stored order of recent files, since it made more
sense - the first is now the most recent.