Fix bugs in prefs/recent files
Preferences were being saved every time AutoImg was
checking for whether to load an image, because the
"docked/floated" flag was set.  Altered so that the
prefs are only saved when the value is changed, not
if it is set to the same value as it formerly was.

When getting a preference that was a list (like the list
of recent files, the code was not doing a `copy` (due to my
Python ignorance). This meant that when the new recent
file was added, it was inserted directly into the prefs list
rather than "setting" the list, hence saving wasn't being
triggered. Fixed by using `deepcopy` - in fact, a shallow
`copy` would be OK for the current settings stored in
prefs, but hopefully a deep copy will avoid similar
headaches in the future. Performance is not an issue
for the relatively small number of prefs values.