Implement page details dialog
This allows user to configure page labels in a similar manner
to in GG1.
1. Pages are listed in order with style of page number set to
Arabic/Roman/ditto, and page number set to increment from
previous or set to specific number or be a "No Count" page.
2. User can click on Arabic/Roman or page number setting
to cycle round possible values.
3. Unlike GG1, user does not need to use Ctrl/Shift clicking,
and labels are recalculated automatically.
4. Cancel discards all changes, and OK uses the new settings.
5. Note, such changes are not saved to disk unless file is saved
(which is why the "modified" flag is set, so user will be prompted
to save on exit).