Implement pptxt without subprocess
Instead of writing to temp file, then spawning process to run
tool which reads temp file and writes to a temp log file, which
GG then reads, uses `io.StringIO` objects as surrogate stdin
and stdout, and passes them directly to the `pptxt` function.

The `pptxt` module can also be run as a standalone tool, either in
a poetry shell, or if `regex` module has been installed in main
python environment.