Add checksum to bin file to trap mismatches
User might have edited main file in an editor other than GG.
This might mean page boundary positions stored in the bin
file no longer match the positions in the text file.
By storing an MD5 checksum in the bin file, check that the
bin file and text file match, and warn if not.