Add pptxt.py
Add the tool. Integration with GG will come via a subsequent PR.

Note that the tool is based on Roger's Go version of pptxt, with
some adaptation, such as not doing spellcheck (which GG does
independently). Also, this version of the tool is able to use
color in its output to highlight words/characters.