Fix short/empty line
1. Remove Columnize Selection menu option - not really needed
because the user can either do a column selection in the first
place, or if they do a regular selection the Column Cut/Copy/Paste
options all columnize it before Cut/Copy/Paste happens.
2. Handle case where a line is so short (or empty) that no text
will be column selected in that line. Since it isn't possible to
tag a zero-length string as selected, we select the newline
(ie to the start of the next line). Then when copying/cutting,
detect that case