Fix macOS shortcut key bindings
Primary problem is to do with case on the Mac. Having
`Ctrl+Shift+S` and `Ctrl+Shift+s` appeared to stop them
from working, whereas that is useful on other platforms
due to behavior when Caps Lock is on. More details in
code comments.

Couple of minor fixes:

When called via shortcut, functions
get called with an event arg, but not when called via
menu. Such functions therefore need optional arg.

Also, after `Cmd+o` and `Cmd+Shift+s` which pop file
selection dialogs, on macOS focus was not returned to
the main window, which can confuse users.