Add column (block) Cut, Copy & Paste (#110)
* Add column (block) Cut, Copy & Paste

1. Add classes to handle Tk Text indexes. GG1 has many
occurrences of combining a row/column to get a string
Tk Text index, or split an index into the row & column.
To reduce this, store in a class which can be constructed
with either form, and from which either form can be
extracted. Also add a class to handle a single range of
such indexes, i.e. a start and end.
2. Add Columnize Selection, which takes the current selection
and replaces it with multiple selections in order to form a
block from the top-left to the bottom-right of the current
selection. This feature may not be required in practice, but
is required internally to do Column Cut/Copy/Paste, so I plan
to leave it in the menu for now.
3. Make Shift+Mouse-click & drag do column selection. Press Shift, then mouse-click, then drag. Should also generally work to click & drag to start normal select, then press Shift and continue dragging to switch to column select - but not vice versa, i.e. starting column selecting then releasing Shift to do normal select is not supported.
4. Add Column Copy (shortcut Shift+Ctrl+C) which either copies
a column selection, or columnizes a normal selection and then
copies it. If there is a column selection, then normal Copy (Ctrl+C)
actually does a Column Copy. Format in clipboard is multiple
lines delimited by newline characters.
5. Add Column Cut (shortcut Shift+Ctrl+X) similar to Column Copy
but cuts instead of copying. Normal Cut does Column Cut if there
is a column selection.
6. Add Column Paste which pastes a previously copied/cut
column selection. If there is no selected text in the main window
the block will be inserted at the insert cursor position. If there is
selected text, then selected lines will be replaced with pasted lines.
If the clipboard text has fewer lines than the selected text, the
clipboard is repeated to fill the space of the selected text. If the
clipboard text has more lines than the selected text, then after
all the selected text lines have been replaced, the remainder of
the clipboard text is pasted in the lines below. Additional lines
and/or spaces are added if necessary to make the column paste
work as a block if near the end of a file or pasting over variable
length lines.

* Minor Column Select improvements

1. Use Option/Alt instead of Shift key
2. Change cursor to cross when doing column select with mouse
3. Fix bug caused by: Hold Shift, Click & drag, Release Shift,
Continue dragging. Tk needs an anchor mark storing in order
for it to cope with returning to non-column select.

* Fix short/empty line

1. Remove Columnize Selection menu option - not really needed
because the user can either do a column selection in the first
place, or if they do a regular selection the Column Cut/Copy/Paste
options all columnize it before Cut/Copy/Paste happens.
2. Handle case where a line is so short (or empty) that no text
will be column selected in that line. Since it isn't possible to
tag a zero-length string as selected, we select the newline
(ie to the start of the next line). Then when copying/cutting,
detect that case