Omit Edit->Preferences menu item on Mac platform
On the Mac platform, tkinter is automatically adding a Settings
menu under the {application_name} menu, making the "Preferences..."
item under the Edit menu redundant.

The customary and "correct" place for the Settings menu item on Mac
is the one which is provided by the framework as a default. This PR
omits the second menu item on the Mac platform only.

Fixes #115