Improve handling of image window
1. Closing a file or loading a new one clears the window. However,
if AutoImg is on, a new image may be loaded.
2. New option "Hide Image" in the View menu closes the image
window (and turns off AutoImg it is was enabled). Closing the
floating image window via the window manager does the same.
3. Dock/Float in View menu replaced with "Dock Image" checkbox.
4. New convenience method `add_checkbox` (like `add_button`)
which takes a label string and processes it to extract the
tilde for keyboard navigation. It accepts an optional shortcut key.
It also takes two callback functions, one for when toggled on,
one for off, as well as the initial value of the boolean.
5. Stopped status bar buttons accepting focus because otherwise
clicking on one (eg "See Img") takes focus from main text widget.