Add `Find All` to Search dialog
Either searches whole file, or just "in selection".

Displayed in a checker dialog. At the moment, this dialog only
has one feature: left-click a line in the checker dialog to jump to
the associated line in the main text window. In future, other
types of click may be added, as in GG1.