Add "Whole word" option to Search
Works by adding a "start of word boundary" at the start, and an
"end of word boundary" at the end.

Unlike in GG1, you can use this with a regex, though if your
regex matches something that does not start/end with a word
character, it may not match what you expect. A good regex
would be `c.t` which would match `cat` and `cot`, but not `cats`.
It is the same as putting `\m` before the regex, and `\M` after.
If it causes confusion, it could be forbidden, but it seems as
though it might be useful, to save users having to remember
the word boundary flags (which are different for Tcl than for
Perl regexes, which use `\b`)