Search (#126)
* Add dummy Search dialog

* Split maintext from mainwindow

The single MainText widget will need to be accessed by many
features, e.g. Search. However, the search feature should not
need to import the whole of mainwindow, which includes menus,
status bar, etc.

* Rudimentary Search feature

1. Dialog popped from Search menu
2. Search forwards (or backwards by Shift clicking Search)
3. Press Enter (or Shift Enter) to Search
4. Regex button enables/disables Tcl-style regexes
5. Wrap button to wrap around end (or start if reversing)
6. Case insensitive - does what it says
7. Reverse - searches backwards (and Shift-Click searches forwards)
8. Count and Find All not implemented yet.

* Pre-populate search box & tidy opposite dir search

* Add search history

1. Extract useful common parts of Combobox into `widgets.py`
2. Combobox can have a Prefs key associated with it, which
causes the contents of the dropdown to persist across runs of
the program.

* Add "Count" feature to Search dialog.

Also, allow Count to be across whole file or just selection.

* Ensure only one of each dialog can be created

Top level dialogs could be instanced multiple times, leading
to several search dialogs being visible, for example.

Instead, provide `widgets.show_toplevel_dialog()` that checks
if the given type of dialog already exists, and just deiconfies
it, if so.

* Add `Find All` to Search dialog

Either searches whole file, or just "in selection".

Displayed in a checker dialog. At the moment, this dialog only
has one feature: left-click a line in the checker dialog to jump to
the associated line in the main text window. In future, other
types of click may be added, as in GG1.

* Stop Shift-click leaving Search button "pressed"

* Add Find Next/Previous

Bound to F3/Shift+F3 (Cmd+G/Shift+Cmd+G on Macs)

Also, Cmd/Ctrl+F put selected text into the Search box of the
dialog, even if the dialog is already popped.

"Wrap" changed to "Wrap around" for clarity.

* Add "Whole word" option to Search

Works by adding a "start of word boundary" at the start, and an
"end of word boundary" at the end.

Unlike in GG1, you can use this with a regex, though if your
regex matches something that does not start/end with a word
character, it may not match what you expect. A good regex
would be `c.t` which would match `cat` and `cot`, but not `cats`.
It is the same as putting `\m` before the regex, and `\M` after.
If it causes confusion, it could be forbidden, but it seems as
though it might be useful, to save users having to remember
the word boundary flags (which are different for Tcl than for
Perl regexes, which use `\b`)

* Ensure only one instance of toplevel dialogs

This commit uses a class method on `ToplevelDialog` class.

* Remove unnecessary passing around of `root`

TIL root is an optional argument to `Toplevel`'s constructor, so
that reduces the need to pass it around.