Store geometry of root and toplevel dialogs (#132)
* Store geometry of root and toplevel dialogs

Currently implemented for the main (root) window and for
dialogs inheriting from ToplevelDialog.

Width & Height and position on screen are stored in the GGprefs
file. If a dialog is set not to be resizable, then the width and height
are ignored when loaded from the GGprefs file, and the dialog is
created at its default size. Currently there are no non-resizable
Toplevel dialogs.