Add Replace to Search dialog
1. Replace replaces string and finds the next occurrence.
2. Replace All replaces all occurrences in file, or just in selection
if `In selection` is checked.
3. Tk Text widget doesn't support regex replacement, so matched
string is "replaced" in the code, then inserted back in file.
4. It remains to be seen if there will be compatibility problems
between Tcl regex and Python regex syntaxes.
5. Note current difference from GG1: to include a capture group in
the replacement string, use `\1`, not `$1`
6. Another difference from GG1 is that a match at the start of a
range (when using Replace All on selection), or a match
immediately after the insert cursor will be found, but will not in GG1.
7. Column selection for "In selection" is not yet supported for
Count, Find All or Replace All