Add ability to set language(s) for a book (#148)
1. Lang button added to status bar
2. User can click button and enter one or more languages,
e.g. "en" or "en+la" or "fr,de" or "en es cz"
3. Languages are stored in the projects bin file.
4. If a file doesn't have a bin file (and when GG first starts) the
language defaults to the most recently used language, which
is stored in the Prefs file.

* Make Title Case Selection language-dependent

List of exception words (such as "a" and "the") should only be
used in English texts.

Also resolved two duplicate keyboard navigation letters in the
Edit menu.