Improve `CheckerDialog.add_entry` method
1. Swap argument order and make the range indicating the
area of interest in the main text an optional argument. This makes
it possible to add general text to the dialog, e.g. header line.
2. If a range is given, automatically add the start of the range as
a prefix to the line of text. Otherwise this code would probably
be duplicated in every tool that wants to add lines to the dialog.
3. Stop using "reserved" word `range` as a variable name.
4. Update Search dialog to use the new checker features, in
particular a header and footer to the Search All results.