Facilitate use of external editor (#146)
Allow user to add "Page Marker Flags" (e.g. `[Pg001]`) before
they save a file and edit it in a different editor.
When file is loaded back into GG2, location of Page Marker
Flags is used to update the page break locations that were
loaded from the bin (.json) file.
Since the bin file and main text file will not match, soften
the consequent error message to a warning if the above
situation is detected.
The "modified" flag on the file is set if any page boundary
locations are updated, because although the text file is
not modified, it needs resaving to save the updated bin file.