Enhance the checker dialog features (#157)
* Enhance the checker dialog features

1. Change underline to colored text in the dialog for highlighting
the word(s) of interest
2. Highlight the chosen line in gray so user knows which one they
clicked on
3. Add right-click to remove error from listing - can also remove
other lines in the dialog (e.g. the header lines) because some
(future) tools have headers partway through their output and a
user  may want to get rid of the header when they have handled
all the errors in that section.
4. Add Ctrl+left-click (Cmd+left-click on Macs) to "process" the
clicked error if possible - see below
5. Add Ctrl+right-click to "process" then remove the error.

Some tools, like jeebies give errors where the correction that needs
making is obvious, so if the user "processes" that error, it can make
the correction for them, swapping he/be.

For testing only (I'll remove before merging) I've put in a dummy
processing function for Spell Check - it just logs an info message
with the range of interest and the error message. This is just to
prove that it gets called with the appropriate mouse clicks. There
will be no "process" function for Spell Check in fact.

Bind to button click, not release, and return "break" to avoid
default behavior. Hopefully will fix right-click on macOS.

Set foreground color on selected checker dialog error, so it is visible
against a light gray background regardless of default text color.

Also, don't force the focus to main text window when user clicks
on error in checker dialog.

Fixed one place in S/R dialog which also forced focus.