Implement "Add word to project dictionary"
1. Small reorganization of files to separate `Root`. Having it in
`mainwindow.py` caused circular imports, because Root is not
part of the main window, and is needed by other dialogs.
2. Used the "process" feature of checker dialog to supply
"Add to project dictionary", either by Ctrl-clicking (Cmd on Mac)
a spelling error message, or by selecting the message then
using the button in the dialog. Other buttons will be added soon.