Refactor: remove `range` as a variable name (#161)
Previously we were using `range` as a variable name or an argument in
some places. `range()` is a global function. This didn't cause any
problems with the Guiguts logic, but it did trip up linters and is not
best practice.

Additionally, `help` was used as a variable in one function. Changed
it for the same reason as above.

At this time pylint is happy (at least, regarding built-in names).

Fixes #137