Destroy & recreate checker dialogs
Checker dialogs are used for a variety of tools, and vary in their
appearance and controls. They therefore need destroying and
recreating, not just deiconifying, if they already exist. This could
be solved by inheriting different dialogs from CheckerDialog,
e.g. SpellCheckDialog, FindAllDialog, etc. However, it would
still be necessary to deal with things like the language changing
or the project dictionary changing while a dialog exists, in which
case some functions in the dialog may not work correctly. This
seems the best fix for now, with very little cost. Other dialogs
that inherit from TopLevelDialog, such as MessageLog, do not
need to be destroyed & re-created.