Checker dialog fixes (#167)
* Destroy & recreate checker dialogs

Checker dialogs are used for a variety of tools, and vary in their
appearance and controls. They therefore need destroying and
recreating, not just deiconifying, if they already exist. This could
be solved by inheriting different dialogs from CheckerDialog,
e.g. SpellCheckDialog, FindAllDialog, etc. However, it would
still be necessary to deal with things like the language changing
or the project dictionary changing while a dialog exists, in which
case some functions in the dialog may not work correctly. This
seems the best fix for now, with very little cost. Other dialogs
that inherit from TopLevelDialog, such as MessageLog, do not
need to be destroyed & re-created.

* Focus & position cursor when CheckerDialog clicked

Position insert cursor at start of current line in CheckerDialog.
This makes it easier to use Ctrl+a, Ctrl+c to copy all messages,
or use Shift+arrow keys to select specific messages for copying.

* Keep "entries" count up to date

When entry is removed from CheckerDialog ,reduce the count.

Also, select the first entry when the dialog is created, and don't
use `see` which scrolls the list unnecessarily.