Fix pptxt/spellcheck re-run confusion
Checker dialogs should be destroyed & re-created rather than
just re-used. When pptxt work was merged, the `destroy` arg
wasn't set to True, so this didn't happen for pptxt, and when
the user tried to re-run pptxt, it re-ran whatever the dialog had
been used for previously.

Make "destroy" True by default for checker dialogs - it is False
by default for other Toplevel dialogs.