Add ordinal & unicode char name to status bar
The current character's hex ordinal, and unicode name are
displayed in a new status bar button.

In future, we may want to re-order the status bar buttons, to put
this last because it changes width by quite large amounts
for different characters. Alternatively, we may decide to make it a
fixed width, or a minimum width that will cover most characters.
Or, we may decide to put some buttons, like language on the
right hand side of the status bar, so they are not affected by the
length of this button.
I'm putting that decision off till later.