Fix bug between Find All and Checker dialogs
If both Find All and another Checker dialog are visible, the
second to be created deletes the marks from the file that show
where the first dialog's messages refer to. Then, clicking on the
first dialog raises an exception because the mark doesn't exist.

Fix by naming the marks after the dialog class, which is always
unique for dialogs that can be shown simultaneously.