Implement click-to-go in WF dialog
Left click to go to word: first time clicking a word will go to the
first occurrence in the file; subsequent clicks will go to the next
occurrence, etc. Clicking on a new word resets to starting from the
beginning of the file again (similar to GG1 behavior, though there
it requires double-click).

Right click to pop Search dialog with the word already entered.

Also, increase pylint's idea of how much code needs to be similar
before reporting it as duplicate from 4 lines to 6. The previous
value is too small - it was reporting something like the following
occurring in 2 places:
```
try:
    my_func()
except ValueError:
    return None
```