Add "selection" label to Status Bar
1. Show range, e.g. "12.5-13.8", for regular selection.
2. Show rows/cols, e.g. "R:3 C:6", for column selection.
3. Click on label to revert to previous selection.

Due to the issue of drag selection consisting of several selections
of increasing length, and we don't want these to fill the previous
selection slot, we only store previous selection if both the start
and end of the selection are different to the previous ones. This
will work in almost every real-life circumstance, and I can't find
a reasonable way to avoid it. I investigated whether it was
possible to detect if the mouse button was pressed (not easily
without interfering with existing click/double-click functionality)
and if it was possible to override every method of doing a
selection (possible, but lots of code to write, attempting to
mimic native widget behavior, which is not written in Python)
and if it was possible to use the `<<Selection>>` event (no,
because of drag selection generating several of these).