Make some focus-forcing behavior OS-specific
Because "deiconify" does not work on (some) X-based systems,
and may not be needed on macOS, make that particular hack
when the program starts specific to Windows.

Also, when clicking on a checker dialog message, macOS users
would expect focus to remain on the checker dialog, but other
users would find it helpful to focus on the main window, so
make that OS-specific too.

Fixes #180