Fix WF marked-up phrase search
Marked up phrases (with "<i>" for example) will not be found if
the wholeword flag is set, because wholeword relies on a boundary
between word and non-word characters, so doesn't match
boundary between "<" and " ".
(word/nonword possibly to be revisited at some point)

So, marked up phrases, and character count matches should
not have the wholeword flag set.