Fix bugs with multiline marked up characters
Both the click to view the location and the right-click to pop the
search dialog failed to convert the Return Arrow character back
into a newline to make the searching work.

Also, the search for non-marked up suspects had the regexp
flag True, which meant that special characters within the string,
such as `|` were interpreted regex-wise, which could lead to
more or fewer matches than expected.