Close dialogs when file is closed or opened (#205)
If any dialogs are popped, close them when user loads a new
text file, or closes the current one.

This work revealed that the marks that are created in the main
text window by checker dialogs like find all, etc., were being left
behind. If there were several thousand of these, then once all the
text was deleted when the file was closed and/or a new one
re-loaded, all those marks ended up on the first line, and
the Text widget choked. Now dialogs can have a reset method
that cleans up after them, and it is called when the dialog is
destroyed.

Fixes #203
Fixes #200