Add sort options, etc to checker dialogs
1. Either sort by line.column or by "Alpha/type", which means
an alphabetic sort where appropriate, or grouped by type of
error.
2. CheckerDialog provides default sort routines, but specific
tools can define their own, like Basic Fixup.
3. Checker messages are now distinguished into header, footer
or content, with different sections - this is so that sorting knows
which ordering to preserve, e.g. a header should come above
content, regardless of sort order.
4. Project dictionary is now only saved if a word was actually
added - previously it was saved even if word was already there.
5. This makes it possible to change the meaning of "process &
remove all" to mean "(process & remove) all", rather than
"process & (remove all)".
6. Previously selected message is reselected and displayed when
results are re-sorted.
7. Some commonizing of code within checkers.py so that one
method now does the work of processing and/or removing,
and is called with various arguments to control it.
8. Add tooltips to checker dialogs, and make tooltips vanish
on button release, not press, so they do not clash
with button presses that actually do something.
9. Also add support for "Shift+Cmd/Ctrl+left button",
which means process this line and all matching lines.