Simplify status bar clicking
1. Avoid using right button. Use left button for main feature, and
Shift-left for additional feature if needed.
2. Move "Choose image scans dir" to File-->Project menu, so no
longer available by clicking status bar. It is popped automatically
anyway, if the user tries to look at a scan before it is set up. GG1
has a File-->Project menu button for it too.
3. Add Shift-click to the ordinal button to toggle the display of
the character name. Some users find the changing length of the
button distracting, and it would take up too much space to make
it long enough to fit the longest common character names. Setting
is stored in GGprefs file.

Fixes #210