Add a Preferences dialog
1. Menu Edit-->Preferences pops the dialog on Windows/Linux,
and the standard application menu Settings entry pops it on a Mac.
2. Currently has 2 tabs (one of which is currently empty)
3. Line numbers, Auto Img, and Bell settings implemented.
4. Existing Preference callback feature utilized and extended to be
called whenever preference changes. This means preference can
be used to set/query the value, rather than needing a separate
variable and/or set/query functions.
5. Should be able to change tabs in the dialog by using Ctrl-tab
(maybe Cmd-tab on Macs?) and by using Alt+first letter of tab name
(probably not on a Mac).