Downgrade invalid PrefKey message
An assertion is too annoying for developers and testers, because
if you run a "later" release, it saves a new PrefKey in the Prefs file.
Then when you run an "earlier" version, it reports that the new
PrefKey it loads from the Prefs file is invalid.

Downgrade it to a warning, which will appear in the terminal,
since it happens before the GUI has been initialized.