Change Page Label Config dialog to modeless
When it's modal, you can't scroll the text up and down to help
decide what values you want to set in the dialog.

Change OkCancelDialog to OkApplyCancelDialog which has an
Apply button that applies the changes made so far in the dialog.
Use this new dialog for Page Label Config.

Note that (like other apps) once you click Apply, those changes
are applied and can't be canceled by clicking Cancel. The OK
button is equivalent to Apply followed by closing dialog.

(Note: nested modal dialogs are not implemented well in
tkinter - when the dialog popped another to ask the page number,
the dialog became modeless once the page number was entered.)