Ensure document has a filename before running tool
1. If tool is run on a document that has never been saved, then it
does not have a "project folder" in which to put files like the
project dictionary. So, check file has a filename before running the
tool.
2. This required something I've been avoiding for a while - making
the single File object accessible from anywhere. Function `the_file`
now returns the single File object (similar to how `maintext` works).
3. Unfortunately file.py imported spell.py in order to reset the
current spelling dictionary, which was the wrong way round, in that
the file shouldn't know about the spell check tool, but came about
because of my previous reluctant to do point 2 above.
4. Now point 2 has been done, the spell checker keeps a record
of which languages it has loaded, and resets if it is asked to
spell check a file with different languages.
5. The structure makes more sense now (at least to me) - there
are just a few objects that are essentially singletons that it is a
pain to pass around everywhere, so having a function that
returns them seems like the least bad option.