Tidy up when dialog is destroyed
For now, just tooltips need to be deleted when a dialog is destroyed,
because the tooltips are Toplevel widgets in their own right, so are
not automatically destroyed when the widget they belong to is
destroyed.

Register each tooltip of a child of a TopLevelDialog, and when that
dialog is destroyed, destroy the tooltips.