Implement Page Separator Fixup
1. Menu entry added to Tools menu.
2. Behavior generally similar to GG1
3. "Auto Fix" option is like "99% Auto" in GG1, but a more
conservative when there are hyphens or hyphen-asterisks
before/after the page separator. These need a human eye.
4. Reason for Undo/Redo buttons is that these need to
normal undo/redo but then to re-view the first page separator
that needs fixing, which a simple Undo/Redo can't do.
5. The "undo block" methods will be replaced by more
general MainText methods at a later date. The idea is that a
single "undo/redo" will undo/redo all the changes made by
the last mouse click.
6. Various key bindings have been set up, to mostly match or
augment those in GG1. In Windows, at least, the buttons have
and underline under the character that will work as a shortcut
when focus is on the dialog (where it remains during the process).
No need for Cmd or Ctrl, just the letter, e.g. "j" for "Join". Some
have an additional shortcut, e.g. the Blank, Section, Chapter ones
also allow you to press the key for the number of blank lines
i.e. 1, 2 or 4. Also, the normal undo & redo shortcuts (like Ctrl+Z)
work as shortcuts to the Undo/Redo buttons in addition to "u"
and "e".
7. The "Auto" setting is stored in the prefs file. Note that with
"Auto Fix", it doesn't start fixing as soon as the dialog is popped,
because users found that disconcerting when GG1 did it. So, just
like GG1, you need to do something in the dialog, typically
click Refresh or the "Auto Fix" radio button, but it could be one
of the manual fix buttons, like Delete, which it will first do, then
continue automatically.