Alternative theming
This uses the standard themes, but then with a couple of added
ones. This method means you can essentially have the non-themed
default, or any of the other (dreadful) ttk themes, or the two
added ones, which are called awdark and awlight.
I think I could download and add other themes like these.