Allow selection in checker dialog with arrow keys
If the focus is on the checker dialog, and up/down arrow is
pressed, it should select the previous/next line in the dialog.

On Macs, selecting a line didn't force focus to the main text
window anyway. This commit also doesn't force focus for
other platforms if the selection was changed by arrow key
(otherwise, you have to refocus on the checker dialog to
hit the arrow key again).

Note that Ctrl+arrow key, etc., do not do the processing,
fixing, deleting all similar, etc., that can be done with the mouse.
This is only for browsing through the errors.