Fix tooltip bugs when Escaping out of a dialog
If the user closes the dialog with the Escape key, there might be tooltip
currently showing. Although the code attempted to destroy any visible
tooltips, it failed to do so if the parent widget had already been
destroyed (NTS: Avoid `suppress`ing Tk exceptions)

Also if the user Escaped out of the dialog in the 1 second gap before
the tooltip was shown, it caused a traceback, again because a destroyed
parent was being referenced.