Trial two-key binding for bookmarks
Use `Cmd/Ctrl-B` then the number of the bookmark to find it.
Use `Shift-Cmd/Ctrl-B` then the number to set it.

Doesn't seem to work well with other dialogs, e.g. if focus is on S&R dialog.
The key binding works, but the second keystroke (the number) gets
inserted into the Search field. I think this is because the `Cmd-B` gets
absorbed by the Search field, and the number then just appears as a number.
No obvious way round that - it appears to be a limitation in the way
multiple key bindings work.

I also don't like the 2 keystrokes, personally. I would much prefer something
like `Ctrl+1` and `Shift+Ctrl+1`.