Tidy/harmonize menu entries
Some menu entries did not have `...` following them.
It appears the most common use of this is when a new
dialog will appear that the user can interact with to
complete the task, and not for just "View xyz" or
"Help About" entries.

Also, put the 4 unmatched markup/bracket/quote
entries into a submenu, to stop the Tools menu becoming
too long.