Make Search/Replace respect page boundaries
Without this, page boundaries all end up at the start
of the replaced text. This commit repositions any
page boundaries that are internal to the replaced
text at the same proportions through the new text
that they were through the old text.

In fact, this is the general text replace function, which
is used internally, not just for Search/Replace.