Fix indexing bugs when using Suspects Only
The matching of the line number in the dialog to the
entry in the list of footnote entries didn't take account
of the Suspects Only setting properly. This led to several
bugs where the wrong item was selected or deleted, or
it linked to the wrong place in the main text.