Simplify use of Tk Variables & Prefs
Previously, when a PersistentBoolean/Int/String (which are
inherited from Tk.Variables) was set, it also set the linked
GG Pref value. This puts the reverse linkage in, which means
that if a separate piece of code updates a Pref, then the
Tk variable will be updated, and any associated checkbox,
etc., will be updated too.

This removes the need to have dialog class variables to
the Pref setting.

Should be no change in behavior.