Preserve geometry of maximized main window (#305)
There is a bug/limitation in how the geometry of the
root window is reported when the window is maximized
(at least on Windows). It reports that the window is the size
of the full screen, but that the top left corner of the window
is wherever it was before maximizing.

This means (like GG1) that when you quit and restart, GG2
starts with its top left somewhere on the screen, but the
size is the full screen size, so some of the window vanishes
off the bottom & right.

Fixed by also saving the "state" flag (which is normal or
zoomed) and restoring that when the program is
restarted. Also, not saving the bad geometry (i.e. big
window but not at top-left) reported after maximizing.