Add font family & size preferences (#300)
Add controls to Prefs dialog to allow user to select
the font family and size. 

Make the default font an empty string and fill it in later
once the fonts can be checked, so that we can make it
DPSM, DVSM, or Courier New.

Add line spacing to Mac

Add a separator to separate "preferred" fonts from the
list of all fonts.

Use the "stash & restore selection" hack on all platforms,
and make it happen `after_idle`, to restore any selection 
the user had when they changed font.