Show image from page label config dialog (#322)
* Show image from page label config dialog

When user clicks on a row in the dialog, jump to the
page in the file, so user can see the text there to aid their
decision about which type of page it is. If user has Auto Img
turned on, the image will show in the image viewer.

Fixes #295

* Show image when selection made via arrow keys

Catch selection event, rather than mouse click event.

* Show current page when dialog is popped

Could be annoying to always pop with first page selected
in dialog, if user has noticed an issue at a specific page.
Since we have to select a page, the current one is most
likely to be correct.