Add page label info to page marker flags
Add the info necessary to construct the labels when in GG1.
Format is `[ImgNNN|Arabic|+1]`
"Arabic" can be "Roman"
"+1" can be "No Count" or a page number

This version must be used in conjunction with a version
of GG1 that has the added info - it won't work with older
versions of GG1, like 1.6.2.