Improve placing of GGprefs on Windows (#348)
* Improve placing of GGprefs on Windows

User's "documents" folder is not necessarily `~/Documents`.
This may be because the user has set its location, but is
most likely because Windows has sneakily placed the
documents on OneDrive and the user has given up
battling against it.

So, interrogate Windows to find the actual location
of the Documents folder (which may not even be called
`Documents`) and use that.

On all platforms, if the GGprefs folder doesn't exist,
create it, and also create any parent folders that are
necessary, e.g. if user's documents folder is
`~/Documents` but that folder doesn't exist, then
create `Documents` first, then the GGPrefs folder
within that. Should never happen, but just in case!
Also, if folders can't be created, or the Prefs file
can't be saved, output an error, rather than crashing
with an exception.

* Fix windows dependency