Restore previous selection after rewrapping
When rewrapping a selection in particular, it was odd that
the selection after rewrapping was not the same piece of
text that was selected beforehand. This was due to
selections being stored as line number/column ranges.

This PR uses marks instead, which stick with the text, so
the text selected afterwards is the same as the text
selected before.

Note that if (for some strange reason) the user rewraps a
column selection, the column selections are extended
to include whole lines, because it doesn't make sense to
rewrap parts of a line.