"Spotlight" instead of select key checker words
When spell check (or other tools) want to highlight a
word in the main text window, use a new tag (spotlight)
instead of the default system selection.
This is more similar to the way GG1 works, and eliminates
problems with having an unwanted selection that
clashes with re-running tools like spell check.