Add "Join footnote to previous" feature
If a footnote is a continuation from a previous page, it will
be marked as a suspect in the Fixup dialog.
When that footnote is selected in the dialog, then
"Join to previous" can be used to join the footnote with
its parent.

At the moment, assumes footnote is well-formatted, and
if it is not, the feature might not work.